chore: expose new modules to restructure code
